requirements.txt
streamlit==1.35.0
psycopg2-binary==2.9.7
pydantic==2.7.0
SQLAlchemy==2.0.18